Move roveapi into the proto dir
